fix(chat): use providers reading instead of request (Issue #5576)
