+++ apps/chat/src/pages/auth/signin/index.tsx
@@ -1,4 +1,4 @@
-import { getProviders, signIn, useSession } from 'next-auth/react';
+import { signIn, useSession } from 'next-auth/react';
 import { useCallback, useEffect, useMemo } from 'react';
 
 import { GetServerSideProps } from 'next';
@@ -182,15 +182,14 @@ export const getServerSideProps: GetServerSideProps = async ({
     };
   }
 
-  const checkProvider = authProviders.some(({ id }) => id === query.provider);
+  const checkProvider = authProviders?.some(({ id }) => id === query.provider);
 
   const providerFromQuery = checkProvider ? query.provider : null;
-  const providers = await getProviders();
   const themesHostDefined = !!process.env.THEMES_CONFIG_HOST;
   return {
     props: {
       provider: DEFAULT_PROVIDER ?? providerFromQuery,
-      providers,
+      providers: authProviders ?? [],
       themesHostDefined,
     },
   };
