+++ apps/chat/src/utils/auth/auth-providers.ts
@@ -1,4 +1,8 @@
-import { OAuthProviderType, Provider } from 'next-auth/providers';
+import {
+  OAuthProviderType,
+  Provider,
+  TokenEndpointHandler,
+} from 'next-auth/providers';
 import Auth0Provider from 'next-auth/providers/auth0';
 import AzureProvider from 'next-auth/providers/azure-ad';
 import AzureB2CProvider from 'next-auth/providers/azure-ad-b2c';
@@ -16,11 +20,35 @@ import {
   providerConfigSchema,
 } from '@/src/types/auth';
 
-import { tokenConfig } from './auth-callbacks';
 import { GitLab } from './custom-gitlab';
+import NextClient from './nextauth-client';
 import PingId from './ping-identity';
 
 const DEFAULT_NAME = 'SSO';
+
+// Need to be set for all providers
+export const tokenConfig: TokenEndpointHandler = {
+  request: async (context) => {
+    let tokens;
+
+    NextClient.setClient(context.client, context.provider);
+
+    if (context.provider.idToken) {
+      tokens = await context.client.callback(
+        context.provider.callbackUrl,
+        context.params,
+        context.checks,
+      );
+    } else {
+      tokens = await context.client.oauthCallback(
+        context.provider.callbackUrl,
+        context.params,
+        context.checks,
+      );
+    }
+    return { tokens };
+  },
+};
 
 const getAzureProvider = (config: ProviderConfig) =>
   config.clientId && config.clientSecret && config.tenantId
